move class to library
not specific to concrete application
